+++ tests/path_utils.py
@@ -1,0 +1,16 @@
+"""Module containing some utilities for testing."""
+
+from pathlib import Path
+import pkg_resources as pkg
+
+__all__ = ["PATH_QMCTORCH", "PATH_TEST", "second_derivative"]
+
+# Environment data
+PATH_QMCTORCH = Path(pkg.resource_filename('qmctorch', ''))
+ROOT = PATH_QMCTORCH.parent
+
+PATH_TEST = ROOT / "tests"
+
+
+def second_derivative(xm1, x0, xp1, eps):
+    return (xm1 - 2 * x0 + xp1) / eps / eps
