+++ tests/test_h2_adf_jacobi.py
@@ -8,8 +8,7 @@ from qmctorch.sampler import Metropolis
 from qmctorch.solver import SolverOrbital
 from qmctorch.scf import Molecule
 from qmctorch.wavefunction import Orbital
-
-from .test_utils import PATH_TEST
+from .path_utils import PATH_TEST
 
 
 class TestH2ADFJacobi(unittest.TestCase):
