+++ tests/test_radial_sto.py
@@ -7,7 +7,7 @@ from qmctorch.utils import (plot_block, plot_data, plot_energy,
                             plot_walkers_traj)
 from qmctorch.scf import Molecule
 from qmctorch.wavefunction import Orbital
-from .test_utils import PATH_TEST, second_derivative
+from .path_utils import PATH_TEST, second_derivative
 
 
 class TestRadialSlater(unittest.TestCase):
